Use class_alias to load php5 or php7 code
Fixes #360
